requirements.txt
requests
matplotlib
seaborn
pandas
numpy
python-dotenv
streamlit